+++ assets/js/wpdk.js
@@ -1140,17 +1140,26 @@ jQuery( function ( $ )
         $( document ).on( WPDKUIComponents.REFRESH_JQUERY_TABS, _initTabs );
 
         // Get tabs
-        var tabs = $( '.wpdk-tabs' );
+        var $tabs = $( '.wpdk-tabs' );
 
-        // Init
-        tabs.tabs();
+        // Get layout
+        var layout = $tabs.data( 'layout' ) || 'horizontal';
+
+        // Build layout
+        if( 'vertical' == layout ) {
+          $tabs.tabs().addClass( "ui-tabs-vertical ui-helper-clearfix" );
+          $tabs.find( 'li').removeClass( "ui-corner-top" ).addClass( "ui-corner-left" );
+        }
+        else {
+          $tabs.tabs();
+        }
 
         if ( document.location.href.indexOf( '#' ) > 0 ) {
           // OoO
         }
         // Enable cookie tabs remember
         else {
-          tabs.each( function ()
+          $tabs.each( function ()
           {
             var id = $( this ).attr( "id" );
             if ( 'undefined' !== typeof(id) ) {