+++ CHANGELOG.md
@@ -20,6 +20,12 @@ For more information on SemVer, please visit http://semver.org/.
 
 ---
 
+## Version 1.7.2
+### 2014-11-??
+
+* Added composer `package.json`
+* Added jQuery tabs vertical support
+
 ## Version 1.7.1
 ### 2014-11-05